% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{aggregate_lhood}
\alias{aggregate_lhood}
\title{Aggregate observations' likelihoods into single posterior}
\usage{
aggregate_lhood(lhoods, hypothesis_vars)
}
\arguments{
\item{lhoods}{data frame with one row per observation and hypothesis
combination, and column \code{log_lhood} with log-likelihood of
observations given hypothesis}

\item{hypothesis_vars}{quoted names of columns for hypothesis variables}
}
\value{
a data frame with one row per combination of hypothesis variables,
  and total log likelihood, log posterior, and posterior probabilities in
  \code{log_lhood}, \code{log_posterior}, and \code{posterior}, respectively
}
\description{
Aggregate observations' likelihoods into single posterior
}

