% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boot_ci}
\alias{boot_ci}
\title{Bootstrapped CIs}
\usage{
boot_ci(data, stats, R = 1000, ci_level = 0.95, h0 = NA, na_rm = TRUE,
  ...)
}
\arguments{
\item{data}{Data that's provided to `boot`}

\item{stats}{Named vector or list of function(d,i) that take data and indices
and return a scalar statistic.}

\item{R}{Number of bootstrap samples (default: 1000)}

\item{ci_level}{Percentage for CI (default 0.95) or lower/upper quantiles of CI.}

\item{h0}{Optional null value of statistics for p value computation}

\item{na_rm}{Whether to remove NAs in quantile calculation}

\item{...}{Additional arguments passed to boot.}
}
\description{
Returns a data_frame with columns stat (names of provided stats), observed
(observed value of stats applied to data, t0 in boot), and ci_lo/ci_hi.
}

