% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ceil_to}
\alias{ceil_to}
\title{Round up to specified ceilings}
\usage{
ceil_to(x, ceilings)
}
\arguments{
\item{x}{number to round}

\item{ceilings}{ceilings to round up to}
}
\value{
the lowest ceiling that is greater than x
}
\description{
Round up to specified ceilings
}

