% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{log_mean_exp}
\alias{log_mean_exp}
\title{Numerically stable mean of logged numbers}
\usage{
log_mean_exp(x)
}
\arguments{
\item{x}{a numeric vector of logged values}
}
\value{
the log of the mean of the exponentiated entries in x.
}
\description{
Use for, e.g., calculating marginal log-likelihood
}

