% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{log_sum_exp}
\alias{log_mean_exp}
\alias{log_sum_exp}
\title{Numerically stable sum of logged numbers}
\usage{
log_sum_exp(x)

log_mean_exp(x)
}
\arguments{
\item{x}{a numeric vector of logged values.}
}
\value{
log of the sum/mean of the exponentiated entries in x.
}
\description{
If any element of x is \code{Inf}, returns \code{Inf}. If all elements are
\code{-Inf}, returns all \code{-Inf}.
}
\section{Functions}{
\itemize{
\item \code{log_mean_exp}: Numerically stable mean of logged numbers
}}

