% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{marginalize}
\alias{marginalize}
\title{Convert joint to marginal probabilities}
\usage{
marginalize(joint, marginal_vars)
}
\arguments{
\item{joint}{data frame with (possibly un-normalized) joint probabilities in
the \code{lhood} column.}

\item{marginal_vars}{quoted names of columns with marginal variables
(retained)}
}
\value{
a data frame with columsn for \code{marginal_vars} and any
  pre-existing grouping, plus marginal log likelihood.
}
\description{
Convert joint to marginal probabilities
}

