% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{normalize_probability}
\alias{normalize_log_probability}
\alias{normalize_probability}
\title{Normalize posterior}
\usage{
normalize_probability(d, prob_var)

normalize_log_probability(d, prob_var)
}
\arguments{
\item{d}{data frame, grouping will be preserved}

\item{prob_var}{quoted name of column with probability to normalize}
}
\value{
A data frame with a column \code{posterior} that sums to 1 within
  each group in \code{d}. Additionally, \code{log_posterior} and
  \code{posterior_choice} columns are created that have the log-posterior and
  TRUE for the maximum within each group/FALSE elsewhere, respectively.
}
\description{
Normalize posterior
}
\section{Functions}{
\itemize{
\item \code{normalize_log_probability}: Operate on log-probability
}}

