% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_val_to_less_than}
\alias{p_val_to_less_than}
\title{Format p-values in "less than" notation}
\usage{
p_val_to_less_than(p, cutoffs = c(0.05, 0.01, 0.001, 1e-04))
}
\arguments{
\item{p}{p-value (vectorized)}

\item{cutoffs}{values to be compared to}
}
\value{
formatted string of the form "p < %f", or "p = %.2f" if p is greater
  than the highest cutoff
}
\description{
Format p-values in "less than" notation
}

