% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_val_to_stars}
\alias{p_val_to_stars}
\title{Format p-values in "stars" notation}
\usage{
p_val_to_stars(p, cutoffs = c(1, 0.05, 0.01, 0.001), stars = c("", "*",
  "**", "***"))
}
\arguments{
\item{p}{p-value (vectorized)}

\item{cutoffs}{numeric vector of p value cutoffs (default: c(1, 0.05, 0.01,
0.001)).}

\item{stars}{characters to use for values less than each cutoff (default:
c("", "*", "**", "***").}
}
\description{
Format p-values in "stars" notation
}
