% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{p_val_to_two_tail}
\alias{p_val_to_two_tail}
\title{Convert a one-tailed to a two-tailed p value}
\usage{
p_val_to_two_tail(ps)
}
\arguments{
\item{ps}{vector of p values}
}
\value{
p values above 0.5 are converted to 2*(1-p), and below to 2*p
}
\description{
Convert a one-tailed to a two-tailed p value
}
