% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{paste_with_last}
\alias{paste_with_last}
\alias{paste_and}
\title{Paste with different separator for last pair}
\usage{
paste_with_last(v, sep, last_sep)

paste_and(v, sep = ", ")
}
\arguments{
\item{v}{vector to paste}

\item{sep}{separator for all but last}

\item{last_sep}{last separator}
}
\description{
Paste with different separator for last pair
}
\section{Functions}{
\itemize{
\item \code{paste_and}: final separator of sep plus "and"
}}

