% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{select_groups}
\alias{select_groups}
\title{Select groups of grouped tbl}
\usage{
select_groups(data, groups, ...)
}
\arguments{
\item{data}{}

\item{groups}{Group numbers to select}
}
\description{
Select groups of grouped tbl
}

