% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{aggregate_lhood}
\alias{aggregate_lhood}
\alias{aggregate_log_lhood}
\title{Aggregate observations' likelihoods into single posterior}
\usage{
aggregate_lhood(lhoods, prob, ...)

aggregate_log_lhood(lhoods, log_prob, ...)
}
\arguments{
\item{lhoods}{data frame with one row per observation and hypothesis
combination, and column \code{log_lhood} with log-likelihood of
observations given hypothesis}

\item{prob, log_prob}{name of column with (log) likelihood values to aggregate}

\item{...}{additional arguments are used to define additional groups to
aggregate likelihood within}
}
\value{
a data frame with one row per group (plus additional grouping vars)
  and total likelihood in \code{prob}
}
\description{
When column \code{x} has the likelihood of one observation under hypotheses
named in column \code{H}
}
\section{Functions}{
\itemize{
\item \code{aggregate_log_lhood}: Aggregate observations' log-likelihood
}}

