% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_val_to_stars}
\alias{p_val_to_stars}
\title{Format p-values in "stars" notation}
\usage{
p_val_to_stars(p, approaching = FALSE)
}
\arguments{
\item{p}{p-value (vectorized)}

\item{approaching}{if FALSE (default), cut off at 0.05. if TRUE, use '.' for
(0.05, 0.10] ('approaching significance')}
}
\description{
Format p-values in "stars" notation
}
