% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_groups}
\alias{select_groups}
\title{Select groups of grouped tbl}
\usage{
select_groups(data, groups)
}
\arguments{
\item{groups}{Group numbers to select}
}
\description{
Useful for examining or testing one example group
}

